Langraph_agent/requirements.txt
langchain-google-genai
langgraph
langchain-core
langchain-community
langchain-experimental
python-dotenv
wikipedia
